Isolate main screens tests
