+++ Corona-Warn-App/src/androidTest/java/testhelpers/TestAppComponent.kt
@@ -4,7 +4,6 @@ import dagger.BindsInstance
 import dagger.Component
 import dagger.android.AndroidInjector
 import dagger.android.support.AndroidSupportInjectionModule
-import de.rki.coronawarnapp.ui.main.MainActivityTestModule
 import testhelpers.viewmodels.MockViewModelModule
 import javax.inject.Singleton
 
@@ -14,7 +13,6 @@ import javax.inject.Singleton
         MockViewModelModule::class,
         FragmentTestModuleRegistrar::class,
         TestAndroidModule::class,
-        MainActivityTestModule::class
     ]
 )
 @Singleton
