+++ Corona-Warn-App/src/debug/AndroidManifest.xml
@@ -17,6 +17,10 @@
 
     <application
         android:requestLegacyExternalStorage="true"
-        tools:node="merge" />
+        tools:node="merge" >
+        <activity
+            android:name="de.rki.coronawarnapp.ui.main.FakeMainActivity"
+            android:theme="@style/AppTheme.Main" />
+    </application>
 
 </manifest>
